install google-cloud-sdk from google distribution rather than apt package; replace miniconda with micromamba
This includes a number of changes to reduce the size of docker images build from this repo.

This installs google-cloud-sdk from google's distribution tarball rather than the apt package for it, in order to save some space. It removes the BigQuery component which is installed by default with google-cloud-sdk (it can be restored later if needed), and also manually removes the install of python3 that comes bundled with google-cloud-sdk since we already have an install at the system level in /bin.

This also replaces miniconda with micromamba as the manager for conda packages. Note that this means we do not have an install of python3 in the base conda environment (further reducing the size of the baseimage), so downstream python usage will require an explicit conda python install in either this image or in downstream images build on top of this one (viral-core). This also also adds symlinks to micromamba from "conda" and "mamba",  allow most commands to make use of micromamba transparently, though there are some breaking differences in API (ex. "conda config --add" vs "micromamba config append").